Started to refactor to a .Standard namespace for the standard game.
